feat: add tiptap renderer
